+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 import useInput from "./useInput";
+import useSignup from "./useSignup";
 
-export { useInput };
+export { useInput, useSignup };
